requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-modeltranslation==0.19.10
django-solo==2.4.0
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
